refactor forecastDiagram part into own controller and fxml
