+++ src/main/resources/sample/weatherapp/weatherApp.fxml
@@ -47,63 +47,8 @@
         <Label fx:id="weatherLabel" text="Content1"/>
       </HBox>
       <HBox>
-        <VBox>
-          <Label fx:id="hourlyForecast" text="hourlyForecast"/>
-          <ScrollPane fx:id="scrollPane" prefWidth="600" prefHeight="400" fitToHeight="true"
-            vbarPolicy="NEVER">
-            <TabPane fx:id="tabPane">
-              <tabs>
-                <Tab text="Temperature" closable="false">
-                  <content>
-                    <LineChart fx:id="tempChart" prefWidth="2000">
-                      <xAxis>
-                        <CategoryAxis fx:id="tempXAxis"/>
-                      </xAxis>
-                      <yAxis>
-                        <NumberAxis fx:id="tempYAxis"/>
-                      </yAxis>
-                    </LineChart>
-                  </content>
-                </Tab>
-                <Tab text="Wind Speed" closable="false">
-                  <content>
-                    <LineChart fx:id="windSpeedChart" prefWidth="2000">
-                      <xAxis>
-                        <CategoryAxis fx:id="windSpeedXAxis" tickLabelRotation="45"/>
-                      </xAxis>
-                      <yAxis>
-                        <NumberAxis fx:id="windSpeedYAxis" />
-                      </yAxis>
-                    </LineChart>
-                  </content>
-                </Tab>
-                <Tab text="Pressure" closable="false">
-                  <content>
-                    <LineChart fx:id="pressureChart" prefWidth="2000">
-                      <xAxis>
-                        <CategoryAxis fx:id="pressureXAxis" tickLabelRotation="45"/>
-                      </xAxis>
-                      <yAxis>
-                        <NumberAxis fx:id="pressureYAxis" />
-                      </yAxis>
-                    </LineChart>
-                  </content>
-                </Tab>
-                <Tab text="Humidity" closable="false">
-                  <content>
-                    <LineChart fx:id="humidityChart" prefWidth="2000">
-                      <xAxis>
-                        <CategoryAxis fx:id="humidityXAxis" tickLabelRotation="45"/>
-                      </xAxis>
-                      <yAxis>
-                        <NumberAxis fx:id="humidityYAxis" />
-                      </yAxis>
-                    </LineChart>
-                  </content>
-                </Tab>
-              </tabs>
-            </TabPane>
-          </ScrollPane>
+        <VBox fx:id="forecastVBox">
+<!--          <fx:include source="ForecastLayout.fxml"/>-->
         </VBox>
         <VBox>
           <Label fx:id="8DayForecast"/>
